Write constant-time variable-base scalar mul generically.
The serial (`u32`/`u64`) implementations use a multiple curve models, passing
between extended and projective coordinates when performing addition and
doubling (respectively). But the AVX2 backend doesn't, so in order to write a
single scalar mult implementation, we have to abstract over the curve models.

To do this, this commit moves the `mult_by_pow_2` function into a `Doubleable`
trait, and renames it (and the other `mult` functions) to `mul` for
consistency.